refactor(api): share JSON-body reading via JsonRequestTrait
TwoFactorController, AccountLifecycleController and UserSessionController each
carried an identical private jsonBody() (graceful decode → assoc array), and
TwoFactorController a stringField() pluck. Extract both into JsonRequestTrait
and use it from the three controllers. No behaviour change.
